# PERF: DatetimeIndex.get_indexer with mismatched tz