Avoid casting to object dtype.

```
In [2]: dti = pd.date_range("2016-01-01", periods=10000, tz="US/Pacific")
In [3]: dti2 = dti.tz_convert("UTC")

In [4]: %timeit dti.get_indexer(dti2)
127 ms ± 4.69 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <-- master
352 µs ± 10.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <-- PR

```