@jbrockmendel 2021-01-24
> do we have any asv's on user facing, e.g. loc with mixed tz's?

i dont think so, this implements one that directly measures get_indexer, with a comment that it reached via `Series.__getitem__`